Move components from pages folder Otherwise, the `next build` is not working
