+++ README.md
@@ -21,6 +21,19 @@ You can find open tasks and issues on [GitHub](https://github.com/lmachens/soc.g
 
 Join us on [Discord](https://th.gl/discord) if you like to contribute.
 
+## Town Graph Improvements
+
+If you are reading this, then you're the perfect candidate to improve the Town
+Graph! Here are some ways in which you could contribute:
+
+- Use terms to translate the graph into other languages.
+- Add share button and sync URL with build.
+- Add building sizes like S/M/L to the data
+  (so that we can calculate needed town size).
+- Calculate needed income for achieving build.
+- Calculate recruitment costs and growth statistics
+  (e.g., how long before you max the stack).
+
 ## Development
 
 First, run the development server:
